+++ CHANGELOG.md
@@ -4,7 +4,9 @@ This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 
 ## main
 
-- REMOVED: Removed deprecated `check_domain_premium_price`. Use `get_domain_prices` instead.
+- REMOVED: Removed deprecated `check_domain_premium_price`. Use `get_domain_prices` instead. (dnsimple/dnsimple-developer#916)
+- REMOVED: Removed deprecated `get_whois_privacy` (dnsimple/dnsimple-developer#919)
+- REMOVED: Removed deprecated `renew_whois_privacy` (dnsimple/dnsimple-developer#919)
 
 ## 4.0.0 - 2025-08-20