+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project uses [Semantic Versioning 2.0.0](http://semver.org/), the format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Fixed
+
+- Fixed `deactivate_dns` using incorrect endpoint type
+
 ## 5.0.0 - 2026-01-23
 
 ### Removed